Added DLL files of Oculus VR Library.
This also fixes an issue that occurred in a SonarQube analysis,
and fixes a minor issue indicated by SonarQube.

The analysis issue could cause an analysis to fail with a message
"Can not add the same measure twice on ...". This has been fixed
by manually providing some parameters (avoid using the
Visual Studio Bootstrapper Plugin). Also, StyleCop analysis fails,
so should be disabled.