Fixed compile / linker errors in server.
The UnitTests subproject now includes the GTest sources, which
is the recommended (and most portable) way to build unit tests.
As such, the server no longer depends on GTEST_HOME to be defined,
as it is compiled on the fly.

This also solves the issue on Mac OS X where the linker complains
about not finding `_main`.