Add non-level marker rotation correction to TransformTest
This is the closest I can get without having an actual Meta, but you can
see that it is very easy to tweak the transformations based on
requirements. The only thing missing is the rotation correction of the
Meta level marker itself, but that works exactly the same as the other
markers.